Fix integer overflow on freqs >2.14 GHz.
This changes all variables that store the tuner center frequency to use
int64_t to avoid overflowing 32 bit signed ints.

It fixes the following error:

$ ./rtl_power_fftw -f 2150M
Invalid frequency given to --freq: -2147483648.
Expecting a positive number, allowing the k,M,G multipliers. Exiting.

[cherry-picked from master -- @alajovic]

+++ src/acquisition.cxx
@@ -232,7 +232,7 @@ Acquisition::Acquisition(const Params& params_,
                          Rtlsdr& rtldev_,
                          Dispatcher& dispatcher_,
                          int actual_samplerate_,
-                         int freq_) :
+                         int64_t freq_) :
   pwr(params_.N), actual_samplerate(actual_samplerate_), params(params_),
   aux(aux_), rtldev(rtldev_), dispatcher(dispatcher_), freq(freq_),
   queue_histogram(params.buffers+1, 0)
